refactor(websearch): extract shared helpers into chunkhound/utils/websearch_core.py
